+++ frontend/src/components/my-proposals-profile-page-handler/profile-with-my-proposals.graphql
@@ -22,4 +22,12 @@ query MyProfileWithSubmissions($conference: String!, $language: String!) {
       }
     }
   }
+
+  conference(code: $conference) {
+    id
+    deadline(type: "cfp") {
+      id
+      status
+    }
+  }
 }
